backend/requirements.txt
elasticsearch
fastapi
uvicorn
python-dotenv
cors
sentence-transformers
PyPDF2
python-multipart
motor
pydantic
pymongo